+++ R/generics.R
@@ -289,7 +289,9 @@ setGeneric("obj_format", function(obj) standardGeneric("obj_format"))
 #' @rdname lab_name
 #' @exportMethod obj_format
 setMethod("obj_format", "ANY", function(obj) attr(obj, "format", exact = TRUE))
-
+#' @rdname lab_name
+#' @export
+setMethod("obj_format", "fmt_config", function(obj) obj@format)
 
 #' @export
 #' @rdname lab_name
@@ -301,6 +303,12 @@ setMethod("obj_format<-", "ANY", function(obj, value) {
   attr(obj, "format") <- value
   obj
 })
+#' @rdname lab_name
+#' @export
+setMethod("obj_format<-", "fmt_config", function(obj, value) {
+  obj@format <- value
+  obj
+})
 
 #' @rdname lab_name
 #' @export
@@ -310,6 +318,10 @@ setGeneric("obj_na_str", function(obj) standardGeneric("obj_na_str"))
 setMethod("obj_na_str", "ANY", function(obj) attr(obj, "format_na_str", exact = TRUE))
 #' @rdname lab_name
 #' @export
+setMethod("obj_na_str", "fmt_config", function(obj) obj@format_na_str)
+
+#' @rdname lab_name
+#' @export
 setGeneric("obj_na_str<-", function(obj, value) standardGeneric("obj_na_str<-"))
 #' @exportMethod obj_na_str<-
 #' @rdname lab_name
@@ -317,7 +329,38 @@ setMethod("obj_na_str<-", "ANY", function(obj, value) {
   attr(obj, "format_na_str") <- value
   obj
 })
+#' @rdname lab_name
+#' @export
+setMethod("obj_na_str<-", "fmt_config", function(obj, value) {
+  obj@format_na_str <- value
+  obj
+})
 
+#' @rdname lab_name
+#' @export
+setGeneric("obj_align", function(obj) standardGeneric("obj_align"))
+#' @rdname lab_name
+#' @exportMethod obj_align
+setMethod("obj_align", "ANY", function(obj) attr(obj, "align", exact = TRUE))
+#' @rdname lab_name
+#' @export
+setMethod("obj_align", "fmt_config", function(obj) obj@align)
+
+#' @rdname lab_name
+#' @export
+setGeneric("obj_align<-", function(obj, value) standardGeneric("obj_align<-"))
+#' @exportMethod obj_align<-
+#' @rdname lab_name
+setMethod("obj_align<-", "ANY", function(obj, value) {
+  attr(obj, "align") <- value
+  obj
+})
+#' @rdname lab_name
+#' @export
+setMethod("obj_align<-", "fmt_config", function(obj, value) {
+  obj@align <- value
+  obj
+})
 
 #' General title/footer accessors
 #'